Refactor HmGui - split hmgui.rs into a separate modules
