+++ engine/lib/phx/src/ui/hmgui/focus.rs
@@ -1,0 +1,13 @@
+#[derive(Copy, Clone, PartialEq, Eq)]
+pub enum FocusStyle {
+    None,
+    Fill,
+    Outline,
+    Underline,
+}
+
+#[derive(Copy, Clone, PartialEq, Eq)]
+pub enum FocusType {
+    Mouse,
+    Scroll,
+}
